secure_file_share/requirements.txt
Flask
python-dotenv
pycryptodome
